Imported 1.2.12
git-svn-id: https://svn.code.sf.net/p/vice-emu/code/trunk@17533 379a1393-f5fb-40a0-bcee-ef074d9b53f7

+++ vice/src/c1541.c
@@ -2203,7 +2203,7 @@ static char *floppy_read_directory(vdrive_t *vdrive, const char *pattern)
         }
         p++;
         line[len++] = '\n';
-        bufcat(outbuf, &outbuf_size, (unsigned int*)&max_outbuf_size, line, len);
+        bufcat(outbuf, &outbuf_size, (size_t *)&max_outbuf_size, line, len);
     }
     vdrive_close(vdrive, 0);
